+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/login/LoginActivity.kt
@@ -6,7 +6,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import com.emmsale.databinding.ActivityLoginBinding
 import com.emmsale.presentation.ui.login.uistate.LoginUiState
-import com.emmsale.presentation.utils.extensions.setContentView
+import com.emmsale.presentation.utils.binding.setContentView
 import com.google.android.material.snackbar.Snackbar
 
 class LoginActivity : AppCompatActivity() {
